server/requirements.txt
quart>=0.18.4
quart-cors>=0.7.0
aiohttp>=3.9.1
sounddevice>=0.4.6
numpy>=1.24.3
scipy>=1.10.1
playsound>=1.3.0
nltk>=3.8.1
groq>=0.18.0
requests==2.31.0
jira
python-dotenv==1.0.1
gunicorn==21.2.0
flask==3.0.2
flask-cors==4.0.0
asgiref
uvicorn
PyAudio==0.2.14

# Testing dependencies
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
responses==0.23.1
